Check if sufficient rejections for overlap plot
When applying the plotMethodsOverlap function to null comparisons or with a very small alpha, sometimes no rejections are found, which throws an error from upsetr that isn't very informative. This change adds a check and informative error message. It requires that at least two methods had rejections at the specified alpha.